requirements.txt
aiogram==3.4.1
aiohttp==3.9.3
python-dotenv==1.0.1
google-generativeai>=0.3.0
psycopg2-binary>=2.9.9
